requirements.txt
# Core AI
groq>=0.9.0
google-genai>=0.8.0

# Memory
qdrant-client>=1.9.0
sentence-transformers>=3.0.0
fastembed>=0.3.0
numpy>=1.26.0

# Search
tavily-python>=0.3.0
duckduckgo-search>=6.0.0

# CLI
prompt-toolkit>=3.0.0

# Config
python-dotenv>=1.0.0